+++ design-system/rating/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/rating
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.2.6
 
 ### Patch Changes